+++ scm-ui/ui-webapp/src/repos/components/form/repositoryValidation.ts
@@ -25,6 +25,11 @@
 import { validation } from "@scm-manager/ui-components";
 
 const nameRegex = /(?!^\.\.$)(?!^\.$)(?!.*[.]git$)(?!.*[\\\[\]])^[A-Za-z0-9\.][A-Za-z0-9\.\-_]*$/;
+const namespaceExceptionsRegex = /^(([0-9]{1,3})|(create))$/;
+
+export const isNamespaceValid = (name: string) => {
+  return nameRegex.test(name) && !namespaceExceptionsRegex.test(name);
+};
 
 export const isNameValid = (name: string) => {
   return nameRegex.test(name);