+++ scm-webapp/src/test/java/sonia/scm/api/v2/resources/RepositoryTestBase.java
@@ -65,6 +65,6 @@ abstract class RepositoryTestBase {
         dtoToRepositoryMapper,
         manager,
         repositoryBasedResourceProvider)),
-      of(repositoryCollectionResource));
+            of(repositoryCollectionResource));
   }
 }